pause dodo

+++ includes/ft_murmurhash.h
@@ -21,6 +21,6 @@
 # include <string.h>
 # include "ft_types.h"
 
-t_uint32	ft_murmurhash2(const char* key, size_t len, const t_uint32 seed);
+t_uint32	ft_murmurhash2(const char* key, int len, const t_uint32 seed);
 
 #endif /* !FT_MURMURHASH2_H */
